requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
pypdf
langchain
langchain-core
langchain-text-splitters
langchain-ollama
langchain-community
faiss-cpu
